# deploy/crm/vantriq-backend/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VantriqAI CRM API",
    "version": "9.13.0",
    "description": "The CRM behind VantriqAI: clients, invoices, payments, quotes and proposals,\ncontracts, statements, and the webhook surface the AI agents write through.\n\n## Authentication\n\nTwo ways in, and one of them is for people:\n\n- **Staff session** — `Authorization: Bearer <token>`, issued by `POST /api/auth/verify`\n  after a password and an emailed one-time code. A staff member's role decides reach.\n- **API key** — `x-api-key: <key>`, for machines. Keys are scoped.\n\n## Scopes\n\nLeast to most: `webhook` < `automation` < `staff` < `admin`. A scope opens everything\nbelow it. The admin key opens everything and is intended as break-glass only — set\n`ALLOW_API_KEY_LOGIN=false` once staff accounts exist.\n\n## Automation writes through this API, never to the database\n\nn8n and any other integration must use these endpoints rather than connecting to\nPostgres. Tenancy, quota and tax rules live here; a direct write bypasses all three."
  },
  "servers": [
    {
      "url": "https://{host}",
      "variables": {
        "host": {
          "default": "crm.vantriqai.com"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "staffSession": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
  "paths": {
...
    "/api/analytics/clients/{id}": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "One customer's analytics — identical to what they see on their portal's Analytics tab.",
        "description": "Requires the **staff** scope or higher. Staff session, or an admin key. Day-to-day CRM work.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
    "/api/analytics/platform": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Every customer's conversations and satisfaction together, with the busiest customers.",
        "description": "Requires the **staff** scope or higher. Staff session, or an admin key. Day-to-day CRM work.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
    "/api/analytics/sales": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Our own leads, wins, losses, funnel, sources and what prospects ask the sales agent.",
        "description": "Requires the **staff** scope or higher. Staff session, or an admin key. Day-to-day CRM work.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
...
    "/api/clients/{id}": {
      "delete": {
        "tags": [
          "clients"
        ],
        "summary": "DELETE /api/clients/:id",
        "description": "Requires the **automation** scope or higher. Automation key or any staff session. Clients, invoices, packages, agents.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      },
      "get": {
        "tags": [
          "clients"
        ],
        "summary": "GET /api/clients/:id",
        "description": "Requires the **automation** scope or higher. Automation key or any staff session. Clients, invoices, packages, agents.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      },
      "put": {
        "tags": [
          "clients"
        ],
        "summary": "PUT /api/clients/:id",
        "description": "Requires the **automation** scope or higher. Automation key or any staff session. Clients, invoices, packages, agents.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or an invalid or revoked key"
          },
          "403": {
            "description": "Authenticated, but not to this scope"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "staffSession": []
          }
        ]
      }
    },
...
  }
}